[Gold V] Title: 용액, Time: 108 ms, Memory: 17120 KB -BaekjoonHub

+++ 백준/Gold/2467. 용액/용액.cs
@@ -1,11 +1,11 @@
 var n = int.Parse(Console.ReadLine());
-var liq = Console.ReadLine().Split().Select(long.Parse).OrderBy(x => x).ToArray();
+var liq = Console.ReadLine().Split().Select(int.Parse).OrderBy(x => x).ToArray();
 int l = 0, r = n-1;
-long min = liq[l] + liq[r];
+int min = liq[l] + liq[r];
 var ans = (liq[l], liq[r]);
 while (l < r)
 {
-    long sum = liq[l] + liq[r];
+    int sum = liq[l] + liq[r];
     if (sum == 0) { ans = (liq[l], liq[r]); break; }
     if (Math.Abs(sum) < Math.Abs(min)) { ans = (liq[l], liq[r]); min = sum; }
     if(sum < 0) { l++; }
